[PRO] Added Player Health Regen after some time & changed all Damage and health to float

+++ Project_Ares/Assets/Code/Game/Weapons/Bullet.cs
@@ -9,7 +9,7 @@ namespace ProjectAres {
 
         [Header("Balancing")]
         [SerializeField] float m_killDistance = 1000;
-        [SerializeField] int m_damage = 1;
+        [SerializeField] float m_damage = 1;
 
         Player m_source = null;
