+++ Project_Ares/Assets/Code/Game/Weapons/Rocket.cs
@@ -13,7 +13,7 @@ namespace ProjectAres {
         [Header("Balancing")]
         [Tooltip("distance form levelorigion to autodestry")]
         [SerializeField] float m_killDistance = 1000;
-        [SerializeField] int m_damage = 1;
+        [SerializeField] float m_damage = 1;
 
         Player m_source;
         Rigidbody2D m_rb;
